build your own lisp, finishing up!
unfortunately, the builtin_load function needs a reference to mpc_parser_t* Lispy in order to work.
That means that having the functionality split up like I had it in earlier chapters is no longer easy.
It can still be done but requires a good chunc of refactoring.

The really unfortunate bit is that the tests I had written no longer work, I'll need to have a look at proper testing frameworks in C and do some refactoring here as well.